feat: created new repository and request contracts
