+++ src/domain/ItunesPodcastCollectionResponseContents.ts
@@ -1,0 +1,13 @@
+interface ItunesPodcastCollectionEntry {
+  id: { attributes: { "im:id": string } };
+  "im:name": { label: string };
+  "im:image": { label: string }[];
+  "im:artist": { label: string };
+  summary: { label: string };
+}
+
+export interface ItunesPodcastCollectionResponseContents {
+  feed: {
+    entry: ItunesPodcastCollectionEntry[];
+  };
+}
