+++ src/domain/ItunesRepository.ts
@@ -1,0 +1,6 @@
+import { Podcast } from "./Podcast";
+
+export interface ItunesRepository {
+  getAll(): Promise<Podcast[]>;
+  getById(id: string): Promise<Podcast>;
+}
